modify file names
