+++ pkg/grpc/dapr/dapr_api_test.go
@@ -20,14 +20,11 @@ import (
 	"context"
 	"fmt"
 	"github.com/dapr/components-contrib/bindings"
-	"github.com/dapr/components-contrib/secretstores"
 	"github.com/dapr/components-contrib/state"
 	"github.com/golang/mock/gomock"
 	"github.com/phayes/freeport"
 	"github.com/stretchr/testify/assert"
 	grpc_api "mosn.io/layotto/pkg/grpc"
-	dapr_common_v1pb "mosn.io/layotto/pkg/grpc/dapr/proto/common/v1"
-	"mosn.io/layotto/pkg/mock/components/secret"
 	mock_state "mosn.io/layotto/pkg/mock/components/state"
 	"net"
 	"testing"
@@ -129,272 +126,4 @@ func createTestClient(port int) *grpc.ClientConn {
 		panic(err)
 	}
 	return conn
-}
-
-func TestNewDaprAPI_GetSecretStores(t *testing.T) {
-	fakeStore := secret.FakeSecretStore{}
-	fakeStores := map[string]secretstores.SecretStore{
-		"store1": fakeStore,
-		"store2": fakeStore,
-		"store3": fakeStore,
-		"store4": fakeStore,
-	}
-
-	expectedResponse := "life is good"
-	storeName := "store1"
-	//deniedStoreName := "store2"
-	restrictedStore := "store3"
-	unrestrictedStore := "store4"     // No configuration defined for the store
-	nonExistingStore := "nonexistent" // Non-existing store
-
-	testCases := []struct {
-		testName         string
-		storeName        string
-		key              string
-		errorExcepted    bool
-		expectedResponse string
-		expectedError    codes.Code
-	}{
-		{
-			testName:         "Good Key from unrestricted store",
-			storeName:        unrestrictedStore,
-			key:              "good-key",
-			errorExcepted:    false,
-			expectedResponse: expectedResponse,
-		},
-		{
-			testName:         "Good Key default access",
-			storeName:        storeName,
-			key:              "good-key",
-			errorExcepted:    false,
-			expectedResponse: expectedResponse,
-		},
-		{
-			testName:         "Good Key restricted store access",
-			storeName:        restrictedStore,
-			key:              "good-key",
-			errorExcepted:    false,
-			expectedResponse: expectedResponse,
-		},
-		//{
-		//	testName:         "Error Key restricted store access",
-		//	storeName:        restrictedStore,
-		//	key:              "error-key",
-		//	errorExcepted:    true,
-		//	expectedResponse: "",
-		//	expectedError:    codes.PermissionDenied,
-		//},
-		//{
-		//	testName:         "Random Key restricted store access",
-		//	storeName:        restrictedStore,
-		//	key:              "random",
-		//	errorExcepted:    true,
-		//	expectedResponse: "",
-		//	expectedError:    codes.PermissionDenied,
-		//},
-		//{
-		//	testName:         "Random Key accessing a store denied access by default",
-		//	storeName:        deniedStoreName,
-		//	key:              "random",
-		//	errorExcepted:    true,
-		//	expectedResponse: "",
-		//	expectedError:    codes.PermissionDenied,
-		//},
-		//{
-		//	testName:         "Random Key accessing a store denied access by default",
-		//	storeName:        deniedStoreName,
-		//	key:              "random",
-		//	errorExcepted:    true,
-		//	expectedResponse: "",
-		//	expectedError:    codes.PermissionDenied,
-		//},
-		{
-			testName:         "Store doesn't exist",
-			storeName:        nonExistingStore,
-			key:              "key",
-			errorExcepted:    true,
-			expectedResponse: "",
-			expectedError:    codes.InvalidArgument,
-		},
-	}
-	// Setup Dapr API server
-	grpcAPI := NewDaprAPI_Alpha(&grpc_api.ApplicationContext{
-		"", nil, nil, nil, nil,
-		nil, nil, nil, nil,
-		nil, fakeStores})
-	err := grpcAPI.Init(nil)
-	if err != nil {
-		t.Errorf("grpcAPI.Init error")
-		return
-	}
-	// test type assertion
-	_, ok := grpcAPI.(dapr_v1pb.DaprServer)
-	if !ok {
-		t.Errorf("Can not cast grpcAPI to DaprServer")
-		return
-	}
-	srv, ok := grpcAPI.(DaprGrpcAPI)
-	if !ok {
-		t.Errorf("Can not cast grpcAPI to DaprServer")
-		return
-	}
-	port, _ := freeport.GetFreePort()
-	server := startDaprServerForTest(port, srv)
-	defer server.Stop()
-
-	clientConn := createTestClient(port)
-	defer clientConn.Close()
-
-	client := dapr_v1pb.NewDaprClient(clientConn)
-	for _, tt := range testCases {
-		t.Run(tt.testName, func(t *testing.T) {
-			request := &dapr_v1pb.GetSecretRequest{
-				StoreName: tt.storeName,
-				Key:       tt.key,
-			}
-			resp, err := client.GetSecret(context.Background(), request)
-
-			if !tt.errorExcepted {
-				assert.NoError(t, err, "Expected no error")
-				assert.Equal(t, resp.Data[tt.key], tt.expectedResponse, "Expected responses to be same")
-			} else {
-				assert.Error(t, err, "Expected error")
-				assert.Equal(t, tt.expectedError, status.Code(err))
-			}
-
-		})
-	}
-}
-
-func TestGetBulkSecret(t *testing.T) {
-	fakeStore := secret.FakeSecretStore{}
-	fakeStores := map[string]secretstores.SecretStore{
-		"store1": fakeStore,
-	}
-
-	expectedResponse := "life is good"
-
-	testCases := []struct {
-		testName         string
-		storeName        string
-		key              string
-		errorExcepted    bool
-		expectedResponse string
-		expectedError    codes.Code
-	}{
-		{
-			testName:         "Good Key from unrestricted store",
-			storeName:        "store1",
-			key:              "good-key",
-			errorExcepted:    false,
-			expectedResponse: expectedResponse,
-		},
-	}
-	// Setup Dapr API server
-	// Setup Dapr API server
-	grpcAPI := NewDaprAPI_Alpha(&grpc_api.ApplicationContext{
-		"", nil, nil, nil, nil,
-		nil, nil, nil, nil,
-		nil, fakeStores})
-	// Run test server
-	err := grpcAPI.Init(nil)
-	if err != nil {
-		t.Errorf("grpcAPI.Init error")
-		return
-	}
-	// test type assertion
-	_, ok := grpcAPI.(dapr_v1pb.DaprServer)
-	if !ok {
-		t.Errorf("Can not cast grpcAPI to DaprServer")
-		return
-	}
-	srv, ok := grpcAPI.(DaprGrpcAPI)
-	if !ok {
-		t.Errorf("Can not cast grpcAPI to DaprServer")
-		return
-	}
-	port, _ := freeport.GetFreePort()
-	server := startDaprServerForTest(port, srv)
-	defer server.Stop()
-
-	clientConn := createTestClient(port)
-	defer clientConn.Close()
-
-	client := dapr_v1pb.NewDaprClient(clientConn)
-
-	for _, tt := range testCases {
-		t.Run(tt.testName, func(t *testing.T) {
-			req := &dapr_v1pb.GetBulkSecretRequest{
-				StoreName: tt.storeName,
-			}
-			resp, err := client.GetBulkSecret(context.Background(), req)
-
-			if !tt.errorExcepted {
-				assert.NoError(t, err, "Expected no error")
-				assert.Equal(t, resp.Data[tt.key].Secrets[tt.key], tt.expectedResponse, "Expected responses to be same")
-			} else {
-				assert.Error(t, err, "Expected error")
-				assert.Equal(t, tt.expectedError, status.Code(err))
-			}
-		})
-	}
-}
-
-func TestStateItem2SetRequest(t *testing.T) {
-	req := StateItem2SetRequest(&dapr_common_v1pb.StateItem{
-		Key:      "",
-		Value:    []byte("v"),
-		Etag:     nil,
-		Metadata: nil,
-		Options: &dapr_common_v1pb.StateOptions{
-			Concurrency: dapr_common_v1pb.StateOptions_CONCURRENCY_UNSPECIFIED,
-			Consistency: dapr_common_v1pb.StateOptions_CONSISTENCY_UNSPECIFIED,
-		},
-	}, "appid||key")
-	assert.Equal(t, req.Key, "appid||key")
-	assert.Equal(t, req.Value, []byte("v"))
-	assert.Nil(t, req.ETag)
-	assert.Equal(t, req.Options.Consistency, "")
-	assert.Equal(t, req.Options.Concurrency, "")
-}
-
-func TestDeleteStateRequest2DeleteRequest(t *testing.T) {
-	t.Run("nil", func(t *testing.T) {
-		req := DeleteStateRequest2DeleteRequest(nil, "")
-		assert.NotNil(t, req)
-	})
-	t.Run("normal", func(t *testing.T) {
-		req := DeleteStateRequest2DeleteRequest(&dapr_v1pb.DeleteStateRequest{
-			StoreName: "redis",
-			Key:       "",
-			Etag:      nil,
-			Options: &dapr_common_v1pb.StateOptions{
-				Concurrency: dapr_common_v1pb.StateOptions_CONCURRENCY_LAST_WRITE,
-				Consistency: dapr_common_v1pb.StateOptions_CONSISTENCY_EVENTUAL,
-			},
-			Metadata: nil,
-		}, "appid||key")
-		assert.Equal(t, req.Key, "appid||key")
-		assert.Nil(t, req.ETag)
-		assert.Equal(t, req.Options.Consistency, "eventual")
-		assert.Equal(t, req.Options.Concurrency, "last-write")
-	})
-}
-
-func TestStateItem2DeleteRequest(t *testing.T) {
-	req := StateItem2DeleteRequest(&dapr_common_v1pb.StateItem{
-		Key:      "",
-		Value:    []byte("v"),
-		Etag:     nil,
-		Metadata: nil,
-		Options: &dapr_common_v1pb.StateOptions{
-			Concurrency: dapr_common_v1pb.StateOptions_CONCURRENCY_LAST_WRITE,
-			Consistency: dapr_common_v1pb.StateOptions_CONSISTENCY_EVENTUAL,
-		},
-	}, "appid||key")
-	assert.Equal(t, req.Key, "appid||key")
-	assert.Nil(t, req.ETag)
-	assert.Nil(t, req.ETag)
-	assert.Equal(t, req.Options.Consistency, "eventual")
-	assert.Equal(t, req.Options.Concurrency, "last-write")
 }
